requirements.txt
python-dotenv~=1.1.0
python-docx~=1.1.2

pydantic~=2.11.4

openai~=1.77.0